refactor: reorganize trigger handling by extracting createTriggerCheck and triggerToEvent into separate modules
