+++ src/engine/engine2/eventBus.ts
@@ -1,39 +1,11 @@
-import {
-  GameState,
-  CardLocationID,
-  LogEntry,
-  cardLocationIdIsEqual,
-  BoardCard,
-} from "./engine2";
+import { GameState, CardLocationID, LogEntry, BoardCard } from "./engine2";
 import * as Commands from "./commands";
+import { AttributeType, Priority } from "../../types/cardTypes";
 import {
-  Ability,
-  AttributeType,
-  Priority,
-  TriggerType,
-} from "../../types/cardTypes";
-import { getTargetCards, getTargetPlayers } from "./targeting";
-import {
-  CardAttributeChangedEvent,
-  CardCrittedEvent,
-  CardFiredEvent,
-  CardItemUsedEvent,
-  CardPerformedBurnEvent,
-  CardPerformedFreezeEvent,
-  CardPerformedHasteEvent,
-  CardPerformedHealEvent,
-  CardPerformedPoisonEvent,
-  CardPerformedRegenEvent,
-  CardPerformedShieldEvent,
-  CardPerformedSlowEvent,
-  CardReloadedEvent,
   GameEndedEvent,
   GameEvent,
-  GameFightStartedEvent,
   GameTickEvent,
-  PlayerAttributeChangedEvent,
   PlayerDiedEvent,
-  PlayerOverhealedEvent,
 } from "./events";
 import { getCardAttribute, getPlayerAttribute } from "./getAttribute";
 import {
@@ -281,340 +253,6 @@ export class EventBus {
   ): void {
     this.listeners.delete(eventClass);
   }
-}
-
-function checkSubjectAndTargetIfDefined(
-  gs: GameState,
-  ability: Ability,
-  checkIfTriggeringID: CardLocationID,
-  eventSourceID: CardLocationID,
-  e: GameEvent,
-  targetType: "player" | "card" | null,
-) {
-  if (!ability.Trigger.Subject) {
-    console.warn(
-      `Ability ${ability.InternalName} has no subject, skipping trigger check`,
-    );
-    return false;
-  }
-  // Check subjects include source card
-  const subjects = getTargetCards(
-    gs,
-    ability.Trigger.Subject,
-    checkIfTriggeringID,
-    e,
-  );
-  // if no subjects is source card return false
-  if (
-    !subjects.some((subject) => cardLocationIdIsEqual(subject, eventSourceID))
-  ) {
-    return false;
-  }
-
-  if (ability.Trigger.Target && targetType) {
-    if (targetType === "player") {
-      // Check if target player is same as ability target
-      const targetPlayers = getTargetPlayers(
-        gs,
-        ability.Trigger.Target,
-        checkIfTriggeringID,
-        e,
-      );
-
-      return targetPlayers.some(
-        (playerIdx) => playerIdx === checkIfTriggeringID.playerIdx,
-      );
-    } else if (targetType === "card") {
-      const targetCards = getTargetCards(
-        gs,
-        ability.Trigger.Target,
-        checkIfTriggeringID,
-        e,
-      );
-
-      return targetCards.some((card) =>
-        cardLocationIdIsEqual(card, eventSourceID),
-      );
-    }
-  }
-  return true;
-}
-
-/**
- * Create a trigger check function for an ability
- */
-export function createTriggerCheck(
-  ability: Ability,
-  locationID: CardLocationID,
-): (gs: GameState, e: GameEvent) => boolean {
-  const triggerType = ability.Trigger.$type || "";
-
-  const checker = (gs: GameState, e: GameEvent): boolean => {
-    switch (triggerType) {
-      case TriggerType.TTriggerOnCardFired: {
-        if (e instanceof CardFiredEvent) {
-          return (
-            e.sourceCardID.playerIdx === locationID.playerIdx &&
-            e.sourceCardID.cardIdx === locationID.cardIdx
-          );
-        }
-        return false;
-      }
-      case TriggerType.TTriggerOnFightStarted: {
-        return e instanceof GameFightStartedEvent;
-      }
-      case TriggerType.TTriggerOnItemUsed: {
-        if (e instanceof CardItemUsedEvent) {
-          return checkSubjectAndTargetIfDefined(
-            gs,
-            ability,
-            locationID,
-            e.sourceCardID,
-            e,
-            null,
-          );
-        }
-        return false;
-      }
-      case TriggerType.TTriggerOnCardCritted: {
-        if (e instanceof CardCrittedEvent) {
-          return checkSubjectAndTargetIfDefined(
-            gs,
-            ability,
-            locationID,
-            e.sourceCardID,
-            e,
-            null,
-          );
-        }
-        return false;
-      }
-      // TODO implement  "PreviousValue": null, "CurrentValue": null, "Source": null, see Ramrod's ability 1
-      case TriggerType.TTriggerOnCardAttributeChanged: {
-        if (e instanceof CardAttributeChangedEvent) {
-          if (!ability.Trigger.Subject) {
-            console.warn(
-              `Ability ${ability.InternalName} has no subject, skipping trigger check`,
-            );
-            return false;
-          }
-          // Check subjects include source card
-          const subjects = getTargetCards(
-            gs,
-            ability.Trigger.Subject,
-            locationID,
-            e,
-          );
-          return subjects.some((subject) => {
-            if (cardLocationIdIsEqual(subject, e.modifiedLocationID)) {
-              if (
-                e.attribute === ability.Trigger.AttributeChanged &&
-                ((ability.Trigger.ChangeType === "Gain" &&
-                  e.newValue > e.oldValue) ||
-                  (ability.Trigger.ChangeType === "Loss" &&
-                    e.newValue < e.oldValue))
-              ) {
-                return true;
-              }
-            }
-            return false;
-          });
-        }
-      }
-      case TriggerType.TTriggerOnFightEnded: {
-        return e instanceof GameEndedEvent;
-      }
-      case TriggerType.TTriggerOnCardPerformedHeal: {
-        if (e instanceof CardPerformedHealEvent) {
-          return checkSubjectAndTargetIfDefined(
-            gs,
-            ability,
-            locationID,
-            e.sourceCardID,
-            e,
-            "player",
-          );
-        }
-        return false;
-      }
-      case TriggerType.TTriggerOnCardPerformedBurn: {
-        if (e instanceof CardPerformedBurnEvent) {
-          return checkSubjectAndTargetIfDefined(
-            gs,
-            ability,
-            locationID,
-            e.sourceCardID,
-            e,
-            "player",
-          );
-        }
-        return false;
-      }
-
-      case TriggerType.TTriggerOnCardPerformedShield: {
-        if (e instanceof CardPerformedShieldEvent) {
-          return checkSubjectAndTargetIfDefined(
-            gs,
-            ability,
-            locationID,
-            e.sourceCardID,
-            e,
-            "player",
-          );
-        }
-        return false;
-      }
-      case TriggerType.TTriggerOnCardPerformedPoison: {
-        if (e instanceof CardPerformedPoisonEvent) {
-          return checkSubjectAndTargetIfDefined(
-            gs,
-            ability,
-            locationID,
-            e.sourceCardID,
-            e,
-            "player",
-          );
-        }
-        return false;
-      }
-
-      case TriggerType.TTriggerOnPlayerAttributeChanged: {
-        if (e instanceof PlayerAttributeChangedEvent) {
-          if (!ability.Trigger.Subject) {
-            console.warn(
-              `Ability ${ability.InternalName} has no subject, skipping trigger check`,
-            );
-            return false;
-          }
-          // Check if subjects include source player
-          const subjects = getTargetPlayers(
-            gs,
-            ability.Trigger.Subject,
-            locationID,
-          );
-          // TODO: implement Trigger.Source
-          // Check if attribute that changed was the one specified in the trigger and change type is the same as the trigger
-          if (
-            e.attribute === ability.Trigger.AttributeChanged &&
-            ((ability.Trigger.ChangeType === "Gain" &&
-              e.newValue > e.oldValue) ||
-              (ability.Trigger.ChangeType === "Loss" &&
-                e.newValue < e.oldValue))
-          ) {
-            return subjects.some((subject) => subject === e.playerIdx);
-          }
-          return false;
-        }
-      }
-
-      case TriggerType.TTriggerOnPlayerDied: {
-        if (e instanceof PlayerDiedEvent) {
-          if (!ability.Trigger.Subject) {
-            console.warn(
-              `Ability ${ability.InternalName} has no subject, skipping trigger check`,
-            );
-            return false;
-          }
-          // Check if subjects include source player
-          const subjects = getTargetPlayers(
-            gs,
-            ability.Trigger.Subject,
-            locationID,
-          );
-          return subjects.some((subject) => subject === e.playerIdx);
-        }
-        return false;
-      }
-
-      case TriggerType.TTriggerOnCardPerformedOverHeal: {
-        if (e instanceof PlayerOverhealedEvent) {
-          return checkSubjectAndTargetIfDefined(
-            gs,
-            ability,
-            locationID,
-            e.sourceCardID,
-            e,
-            "player",
-          );
-        }
-      }
-      case TriggerType.TTriggerOnCardPerformedRegen: {
-        if (e instanceof CardPerformedRegenEvent) {
-          return checkSubjectAndTargetIfDefined(
-            gs,
-            ability,
-            locationID,
-            e.sourceCardID,
-            e,
-            "player",
-          );
-        }
-      }
-
-      case TriggerType.TTriggerOnCardReloaded: {
-        if (e instanceof CardReloadedEvent) {
-          return checkSubjectAndTargetIfDefined(
-            gs,
-            ability,
-            locationID,
-            e.sourceCardID,
-            e,
-            "card",
-          );
-        }
-        return false;
-      }
-      case TriggerType.TTriggerOnCardPerformedHaste: {
-        if (e instanceof CardPerformedHasteEvent) {
-          return checkSubjectAndTargetIfDefined(
-            gs,
-            ability,
-            locationID,
-            e.sourceCardID,
-            e,
-            "card",
-          );
-        }
-        return false;
-      }
-      case TriggerType.TTriggerOnCardPerformedSlow: {
-        if (e instanceof CardPerformedSlowEvent) {
-          return checkSubjectAndTargetIfDefined(
-            gs,
-            ability,
-            locationID,
-            e.sourceCardID,
-            e,
-            "card",
-          );
-        }
-        return false;
-      }
-      case TriggerType.TTriggerOnCardPerformedFreeze: {
-        if (e instanceof CardPerformedFreezeEvent) {
-          return checkSubjectAndTargetIfDefined(
-            gs,
-            ability,
-            locationID,
-            e.sourceCardID,
-            e,
-            "card",
-          );
-        }
-        return false;
-      }
-      default: {
-        console.warn(
-          `Unhandled trigger type: ${triggerType} for ability ${ability.InternalName}`,
-        );
-        throw new Error(
-          `Unhandled trigger type: ${triggerType} for ability ${ability.InternalName}`,
-        );
-      }
-    }
-  };
-
-  return checker;
 }
 
 /**
